requirements.txt
pandas
scikit-learn
azureml-core
azureml-pipeline
azure.cli.core